+++ src/components/AppView.js
@@ -1,0 +1,14 @@
+import React, {Component} from 'react'
+
+class AppView extends Component {
+  render() {
+    return (
+      <div>
+        <button onClick={this.props.logOutUser}>Log Out</button>
+        App View
+      </div>
+    )
+  }
+}
+
+export default AppView
